requirements.txt
streamlit
google-genai
openai-whisper
pydantic
ffmpeg-python